Create metareg1.0.stan
This version adds a generated quantities block and moves RR to there, and also adds to new outcomes:

predicted_mean (i.e predicted mean value daily acquisition probability of hand hygiene for 0 to 20 handwashes per day  - with 0 mask use) 
and predicted_mean_new_trial which samples values for intercept and slop from distributions, and gives predictions for a new trial. 
predicted_mean_new_trial is the most relevant one to plot on the graph showing values for the different trials (fig 3 in the current paper).